feat: adjust sick-pigs
